Move functions to 'kudo' package
